IOT-762 - Deaktivering af sletning hvis element er i brug og informering om dette

+++ src/app/profiles/device-profiles/device-profiles-list/device-profiles-list.component.ts
@@ -4,9 +4,10 @@ import { Subscription } from 'rxjs';
 import { DeviceProfile } from '../device-profile.model';
 import { Router, ActivatedRoute } from '@angular/router';
 import { DeviceProfileService } from '../device-profile.service';
-import { HttpErrorResponse } from '@angular/common/http';
 import { MeService } from '@shared/services/me.service';
 import { SharedVariableService } from '@shared/shared-variable/shared-variable.service';
+import { DeleteDialogService } from '@shared/components/delete-dialog/delete-dialog.service';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-device-profiles-list',
@@ -20,16 +21,25 @@ export class DeviceProfilesListComponent implements OnInit, OnDestroy {
   public pageOffset: 0;
 
   public errorMessages: any;
+  deleteDialogSubscription: Subscription;
+  errorTitle: string;
 
   constructor(
     private router: Router,
     private route: ActivatedRoute,
     private deviceProfileService: DeviceProfileService,
     private meService: MeService,
-    private sharedVariableService: SharedVariableService
+    private sharedVariableService: SharedVariableService,
+    private deleteDialogService: DeleteDialogService,
+    private translateService: TranslateService
   ) { }
 
   ngOnInit() {
+    this.translateService
+      .get(['PROFILES.DELETE-FAILED'])
+      .subscribe((translations) => {
+        this.errorTitle = translations['PROFILES.DELETE-FAILED'];
+      });
     this.getDeviceProfiles();
   }
 
@@ -59,14 +69,32 @@ export class DeviceProfilesListComponent implements OnInit, OnDestroy {
 
   deleteDeviceProfile(id: string) {
     if (id) {
-      this.deviceProfileService.delete(id).subscribe((response) => {
-        console.log(response);
-        if (response.ok) {
-          this.getDeviceProfiles();
-        } else {
-          this.showError(response);
+      this.deleteDialogSubscription = this.deleteDialogService.showSimpleDialog().subscribe(
+        (response) => {
+          if (response) {
+            this.deviceProfileService.delete(id).subscribe((response) => {
+              console.log(response);
+              if (response.ok) {
+                this.getDeviceProfiles();
+              } else {
+                if (response?.error?.message === 'Internal server error') {
+                  this.errorMessages = 'Internal server error';
+                  return;
+                } else {
+                  this.deleteDialogSubscription = this.deleteDialogService.showSimpleDialog(
+                    response.error.message,
+                    false,
+                    false,
+                    true,
+                    this.errorTitle).subscribe();
+                }
+              }
+            });
+          } else {
+            console.log(response);
+          }
         }
-      });
+      );
     }
   }
 
@@ -74,16 +102,8 @@ export class DeviceProfilesListComponent implements OnInit, OnDestroy {
     if (this.subscription) {
       this.subscription.unsubscribe();
     }
-  }
-
-  private showError(err: HttpErrorResponse) {
-    if (err?.error?.message == 'Internal server error') {
-      this.errorMessages = 'Internal server error';
-      return;
-    }
-
-    if (err.error?.message) {
-      this.errorMessages = err.error?.message;
+    if (this.deleteDialogSubscription) {
+      this.deleteDialogSubscription.unsubscribe();
     }
   }
 }
